pyproject.toml
[project]
name = "stock-llm"
version = "0.1.0"
description = "Taiwan stock recommendation system with LLM integration"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=0.3.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.2.0",
    "duckdb>=1.0.0",
    "requests>=2.31.0",
    "streamlit>=1.30.0",
    "yfinance>=0.2.40",
    "plotly>=5.24.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/stock_llm"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
